requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-dropbox-storage==0.4.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
drf-yasg==1.17.1
dropbox==10.1.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pathvalidate==2.3.0
Pillow==7.1.1
pipenv==2018.11.26
psycopg2==2.8.5
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
whitenoise==5.0.1
zipp==3.1.0
